# src/Domivice.Users.Web/wwwroot/openapi-original.json
{
    "openapi": "3.0.2",
    "info": {
        "contact": {
            "name": "Domivice Development Team",
            "url": "https://development.domivice.com"
        },
        "description": "The users API",
        "title": "Users API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
    "paths": {
...
        "/v1/users/{userId}/addresses": {
            "get": {
                "operationId": "list_addresses",
                "parameters": [
                    {
                        "description": "The string guid representation of the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }, {
                        "description": "Filter by the address type",
                        "explode": true,
                        "in": "query",
                        "name": "type",
                        "required": false,
                        "schema": {
                            "enum": ["Primary", "Secondary"],
                            "type": "string"
                        },
                        "style": "form"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddressList"
                                }
                            }
                        },
                        "description": "Response when a list of 0 or more addresses is returned"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    }
                },
                "summary": "List addresses",
                "tags": ["Addresses"]
            },
            "post": {
                "operationId": "create_address",
                "parameters": [
                    {
                        "description": "The string guid representation of the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddressCreate"
                            }
                        }
                    },
                    "description": "The json representation of new address",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Address"
                                }
                            }
                        },
                        "description": "Response when new address is created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "There was an error with the request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    }
                },
                "summary": "Create address",
                "tags": ["Addresses"]
            }
        },
        "/v1/users/{userId}/addresses/{addressId}": {
            "delete": {
                "operationId": "delete_address",
                "parameters": [
                    {
                        "description": "The string guid representing the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }, {
                        "description": "The string guid representing the address id",
                        "explode": false,
                        "in": "path",
                        "name": "addressId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Response when the address was deleted successfully"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not allowed to perform action"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The resource was not found"
                    }
                },
                "summary": "Delete Address",
                "tags": ["Addresses"]
            },
            "get": {
                "operationId": "get_address",
                "parameters": [
                    {
                        "description": "The string guid representing the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }, {
                        "description": "The string guid representing the address id",
                        "explode": false,
                        "in": "path",
                        "name": "addressId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Address"
                                }
                            }
                        },
                        "description": "Response when an address is returned"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The resource was not found"
                    }
                },
                "summary": "Get Address",
                "tags": ["Addresses"]
            },
            "patch": {
                "operationId": "update_address",
                "parameters": [
                    {
                        "description": "The string guid representing the user id",
                        "explode": false,
                        "in": "path",
                        "name": "userId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }, {
                        "description": "The string guid representing the address id",
                        "explode": false,
                        "in": "path",
                        "name": "addressId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "style": "simple"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddressUpdate"
                            }
                        }
                    },
                    "description": "The json representation of the address update",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Address"
                                }
                            }
                        },
                        "description": "Response when the address was updated successfully"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "There was an error with the request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not authenticated"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "User is not allowed to perform action"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        },
                        "description": "The resource was not found"
                    }
                },
                "summary": "Patch address",
                "tags": ["Addresses"]
            }
        },
...
    },
    "components": {
...
        "schemas": {
            "Error": {
                "description": "",
                "properties": {
                    "detail": {
                        "description": "A human-readable explanation of the error",
                        "type": "string"
                    },
                    "instance": {
                        "description": "A URI that identifies the specific occurrence of the error",
                        "type": "string"
                    },
                    "status": {
                        "description": "The HTTP response code",
                        "type": "string"
                    },
                    "title": {
                        "description": "A brief, human-readable message about the error",
                        "type": "string"
                    },
                    "type": {
                        "description": "A URI identifier that categorizes the error",
                        "type": "string"
                    },
                    "errors": {
                        "description": "The validation errors",
                        "items": {
                            "$ref": "#/components/schemas/KeyValuePair"
                        },
                        "type": "array"
                    },
                    "extensions": {
                        "description": "The error extensions",
                        "items": {
                            "$ref": "#/components/schemas/KeyValuePair"
                        },
                        "type": "array"
                    }
                },
                "required": ["detail", "instance", "title", "type"],
                "type": "object"
            },
            "KeyValuePair": {
                "description": "Carries key values pair",
                "type": "object"
            },
            "Address": {
                "description": "",
                "example": {
                    "country": "country",
                    "city": "city",
                    "postalCode": "postalCode",
                    "addressLine1": "addressLine1",
                    "addressLine2": "addressLine2",
                    "addressLine3": "addressLine3",
                    "state": "state",
                    "id": "id",
                    "type": "Primary",
                    "userId": "userId"
                },
                "properties": {
                    "userId": {
                        "description": "The user id",
                        "type": "string"
                    },
                    "addressLine1": {
                        "description": "The address line 1",
                        "type": "string"
                    },
                    "addressLine2": {
                        "description": "The address line 2",
                        "type": "string"
                    },
                    "addressLine3": {
                        "description": "The address line 3",
                        "type": "string"
                    },
                    "city": {
                        "description": "The city",
                        "type": "string"
                    },
                    "country": {
                        "description": "The country",
                        "type": "string"
                    },
                    "postalCode": {
                        "description": "The postal code",
                        "type": "string"
                    },
                    "state": {
                        "description": "The state or province ",
                        "type": "string"
                    },
                    "id": {
                        "description": "The address id",
                        "type": "string"
                    },
                    "type": {
                        "description": "The address type. Newly created addresses should default to secondary",
                        "enum": ["Primary", "Secondary"],
                        "type": "string"
                    }
                },
                "required": ["addressLine1", "city", "country", "id", "postalCode", "state", "type", "userId"],
                "type": "object"
            },
...
            "AddressCreate": {
                "description": "",
                "example": {
                    "addressLine1": "some text",
                    "addressLine2": "some text",
                    "addressLine3": "some text",
                    "city": "some text",
                    "country": "some text",
                    "postalCode": "some text",
                    "state": "some text",
                    "type": "primary"
                },
                "properties": {
                    "addressLine1": {
                        "description": "The address line 1",
                        "type": "string"
                    },
                    "addressLine2": {
                        "description": "The address line 2",
                        "type": "string"
                    },
                    "addressLine3": {
                        "description": "The address line 3",
                        "type": "string"
                    },
                    "city": {
                        "description": "The city",
                        "type": "string"
                    },
                    "country": {
                        "description": "The country",
                        "type": "string"
                    },
                    "postalCode": {
                        "description": "The postal code",
                        "type": "string"
                    },
                    "state": {
                        "description": "The state or province ",
                        "type": "string"
                    },
                    "type": {
                        "description": "The address type. Newly created addresses should default to secondary",
                        "type": "string"
                    }
                },
                "required": ["addressLine1", "city", "country", "postalCode", "state", "type"],
                "type": "object"
            },
...
            "AddressList": {
                "description": "Carries a paginated list of addresses",
                "example": {
                    "totalItemsCount": 6,
                    "pageCount": 0,
                    "previousPage": "previousPage",
                    "data": [
                        {
                            "country": "country",
                            "city": "city",
                            "postalCode": "postalCode",
                            "addressLine1": "addressLine1",
                            "addressLine2": "addressLine2",
                            "addressLine3": "addressLine3",
                            "state": "state",
                            "id": "id",
                            "type": "Primary",
                            "userId": "userId"
                        }, {
                            "country": "country",
                            "city": "city",
                            "postalCode": "postalCode",
                            "addressLine1": "addressLine1",
                            "addressLine2": "addressLine2",
                            "addressLine3": "addressLine3",
                            "state": "state",
                            "id": "id",
                            "type": "Primary",
                            "userId": "userId"
                        }
                    ],
                    "nextPage": "nextPage"
                },
                "properties": {
                    "data": {
                        "description": "The address list",
                        "items": {
                            "$ref": "#/components/schemas/Address"
                        },
                        "type": "array"
                    },
                    "nextPage": {
                        "description": "The URL to access the next page",
                        "type": "string"
                    },
                    "pageCount": {
                        "description": "The page count",
                        "type": "integer"
                    },
                    "previousPage": {
                        "description": "The URL to access the previous page",
                        "type": "string"
                    },
                    "totalItemsCount": {
                        "description": "The total items count",
                        "type": "integer"
                    }
                },
                "required": ["data", "pageCount", "totalItemsCount"],
                "type": "object"
            },
...
        }
    }
}